Use variable-length array instead of malloc/free.
The only use of `malloc`/`free` in tomfastmath is in
src/numtheory/fp_prime_random_ex.c, and the size required can be
calculated from the function arguments, so we can just use
a variable-length automatic array instead.

Variable-length automatic arrays are a C99 feature, and TFM already doesn’t
compile as C90, so we might as well take advantage of it.